Organize components into pages and ui folders
