+++ src/App.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import AppHeader from './components/AppHeader.vue'
+import AppHeader from "./components/ui/AppHeader.vue";
 </script>
 
 <template>
