+++ src/components/pages/Sort.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+import { useI18n } from "vue-i18n";
+const { t } = useI18n();
+</script>
+
+<template>
+  <div class="view">
+    <h1>{{ t("sort.title") }}</h1>
+  </div>
+</template>
