+++ src/components/ui/DeleteConfirmModal.vue
